+++ test/community/training/easy/rotatingArrows/CGTest.js
@@ -1,0 +1,64 @@
+import { assert } from 'chai';
+import sinon from 'sinon';
+import File from '../../../../File.js';
+import { assertOutputAnswer } from '../../../../assertOutputAnswer.js';
+import { execute } from '../../../../../lib/community/training/easy/rotatingArrows/rotatingArrows.js';
+
+const __dirname = new URL('.', import.meta.url).pathname;
+
+suite("Rotating arrows", function() {
+    const sandbox = sinon.createSandbox();
+
+    setup(function () {
+        sandbox.stub(console, "log");
+    });
+
+    teardown(function () {
+        sandbox.restore();
+    });
+
+
+    test("Simple", function() {
+        let inputFile = new File(__dirname + 'input/01 - simple.txt');
+
+        execute(inputFile.readline.bind(inputFile));
+
+        assert.strictEqual(
+            console.log.getCall(0).args[0],
+            2
+        );
+    });
+
+    test("More rotate", function() {
+        let inputFile = new File(__dirname + 'input/02 - more rotate.txt');
+
+        execute(inputFile.readline.bind(inputFile));
+
+        assert.strictEqual(
+            console.log.getCall(0).args[0],
+            8
+        );
+    });
+
+    test("Overflow", function() {
+        let inputFile = new File(__dirname + 'input/03 - overflow.txt');
+
+        execute(inputFile.readline.bind(inputFile));
+
+        assert.strictEqual(
+            console.log.getCall(0).args[0],
+            3
+        );
+    });
+
+    test("Repeat", function() {
+        let inputFile = new File(__dirname + 'input/04 - repeat.txt');
+
+        execute(inputFile.readline.bind(inputFile));
+
+        assert.strictEqual(
+            console.log.getCall(0).args[0],
+            18
+        );
+    });
+});