+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Dead men's shot".
 - Tests for "Darts".
 - Tests for "Add'em up".
+- Tests for "Rotating arrows".
 
 ### Fixed
 - File import for "A child's play".